pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dupefinder"
dynamic = ["version"]
description = "A comprehensive tool for finding duplicate and visually similar images"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Zach Daly", email = "zach@example.com"}
]
keywords = ["duplicate", "image", "finder", "photo", "dedup", "hash", "perceptual", "lsh"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]

dependencies = [
    "Pillow>=9.0.0",
    "imagehash>=4.0.0",
    "flask>=2.0.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
progress = [
    "tqdm>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Zedidence/DupeFinderGUI"
"Bug Tracker" = "https://github.com/Zedidence/DupeFinderGUI/issues"
Documentation = "https://github.com/Zedidence/DupeFinderGUI#readme"
"Source Code" = "https://github.com/Zedidence/DupeFinderGUI"

[project.scripts]
dupefinder = "dupefinder.__main__:main"
dupefinder-cli = "dupefinder.cli:main"
dupefinder-gui = "dupefinder.app:main"

[tool.setuptools.dynamic]
version = {attr = "dupefinder.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["dupefinder*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
dupefinder = ["templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--color=yes",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
